requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
billiard==4.2.1
boto3==1.37.18
botocore==1.37.18
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
jmespath==1.0.1
jwt==1.3.1
kombu==5.5.0
motor==3.7.0
passlib==1.7.4
prompt_toolkit==3.0.50
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
redis==5.2.1
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
